Move PGUtil to util folder
