+++ src/Npgsql/BackendMessages/CopyMessages.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Npgsql.Util;
 
 namespace Npgsql.BackendMessages
 {
