+++ src/Npgsql/FrontendMessages/PregeneratedMessage.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Linq;
 using System.Text;
+using Npgsql.Util;
 
 namespace Npgsql.FrontendMessages
 {
