+++ src/Npgsql/INpgsqlDatabaseInfoFactory.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using JetBrains.Annotations;
+using Npgsql.Util;
 
 namespace Npgsql
 {
