+++ src/Npgsql/NpgsqlParameterCollection.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Data.Common;
 using JetBrains.Annotations;
+using Npgsql.Util;
 using NpgsqlTypes;
 
 namespace Npgsql
