+++ src/Npgsql/PgPassFile.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using JetBrains.Annotations;
+using Npgsql.Util;
 
 namespace Npgsql
 {
