+++ test/Npgsql.Benchmarks/TypeHandlers/TypeHandlerBenchmarks.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using Npgsql.Util;
 
 namespace Npgsql.Benchmarks.TypeHandlers
 {
