+++ test/Npgsql.Tests/WriteBufferTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using Npgsql.Util;
 using NUnit.Framework;
 
 namespace Npgsql.Tests
